added user authorization checks

+++ controllers/userController.js
@@ -14,8 +14,8 @@ exports.userLogin = async function (req, res) {
       req.session.loggedIn = true;
       req.session.username = req.body.username;
       res.send({
-        username: user.username, 
-        nickname: user.nickname
+        username: user.username,
+        nickname: user.nickname,
       });
     } else {
       res.status(401).end();
@@ -28,21 +28,27 @@ exports.userLogin = async function (req, res) {
 
 exports.userPOST = async function (req, res) {
   try {
-    const newUser = new User({
-      username: req.body.username,
-      firstName: req.body.firstName,
-      lastName: req.body.lastName,
-      height: req.body.height,
-      weight: req.body.weight,
-      nickname: req.body.nickname,
-      gender: req.body.gender,
-      birthday: new Date(req.body.birthday),
-      entries: [],
-    });
-    newUser.setPassword(req.body.password);
-    await newUser.save();
-
-    res.status(201).end();
+    if (req.session.loggedIn && req.session.username === req.params.username) {
+      const newUser = new User({
+        username: req.body.username,
+        firstName: req.body.firstName,
+        lastName: req.body.lastName,
+        height: req.body.height,
+        weight: req.body.weight,
+        nickname: req.body.nickname,
+        gender: req.body.gender,
+        birthday: new Date(req.body.birthday),
+        entries: [],
+      });
+      
+      newUser.setPassword(req.body.password);
+      
+      await newUser.save();
+      
+      res.status(201).end();
+    } else {
+      res.status(401).end();
+    }
   } catch (error) {
     console.log(error);
     res.status(500).end();
@@ -51,8 +57,12 @@ exports.userPOST = async function (req, res) {
 
 exports.userGET = async function (req, res, next) {
   try {
-    const user = await User.findOne({ username: req.params.username });
-    res.send(user);
+    if (req.session.loggedIn && req.session.username === req.params.username) {
+      const user = await User.findOne({ username: req.params.username });
+      res.send(user);
+    } else {
+      res.status(401).end();
+    }
   } catch (error) {
     console.log(error);
     res.status(500).end();
@@ -61,11 +71,15 @@ exports.userGET = async function (req, res, next) {
 
 exports.userPUT = async function (req, res) {
   try {
-    await User.findOneAndUpdate(
-      { username: req.params.username },
-      { ...req.body }
-    );
-    res.status(200).end();
+    if (req.session.loggedIn && req.session.username === req.params.username) {
+      await User.findOneAndUpdate(
+        { username: req.params.username },
+        { ...req.body }
+      );
+      res.status(200).end();
+    } else {
+      res.status(401).end();
+    }
   } catch (error) {
     console.log(error);
     res.status(500).end();
@@ -74,15 +88,18 @@ exports.userPUT = async function (req, res) {
 
 exports.userDELETE = async function (req, res) {
   try {
-    const deletedUser = await User.findOneAndDelete({
-      username: req.params.username,
-    });
-
-    console.log(deletedUser);
-
-    await Entry.deleteMany({ _id: { $in: deletedUser.entries } });
-
-    res.status(200).end();
+    if (req.session.loggedIn && req.session.username === req.params.username) {
+      // delete user
+      const deletedUser = await User.findOneAndDelete({
+        username: req.params.username,
+      });
+      // delete entries of user
+      await Entry.deleteMany({ _id: { $in: deletedUser.entries } });
+  
+      res.status(200).end();
+    } else {
+      res.status(401).end();
+    }
   } catch (error) {
     console.log(error);
     res.status(500).end();
